Complete create a Dockerfile and test it locally.
This is 3 of 7 parts of the end-to-end application that includes a
CI pipeline.

In order to expose the Flask port in a Docker container, we have to
modify main.py by adding the parameter host="0.0.0.0" when calling
app.run().

Add commands docker_clean and docker_run to Makefile.

Next part is to require status checks to pass before merging PR.