Complete require status checks to pass before merging PR
This is 4 of 7 parts of the end-to-end application that includes a
CI pipeline.

We use the python package prospector to perform static analysis.

Each time you make a PR for this repo, CircleCI will automatically
run your config.yml file. If any steps fail, then you. will not be able
to merge your PR with the main branch.

Add docker-compose.yml file that performs the static analysis for
our custom image archiveso:latest. We need to ensure that the
"make docker_build" command gets executed before it in our CI
pipeline.

Next part is to build, test, tag and upload our web app image using CI.