Complete build, test, tag and upload our web app image using CI (#3)
Complete build, test, tag and upload our web app image using CI

This is 5 of 7 parts of the end-to-end application that includes a
CI pipeline.

We use a workflow to orchestrate a sequential job execution with
dependency. If the static-analysis job passes, we immediately run
deploy-image job.

Jobs in a workflow are isolated from each other. For the deploy-
image job, we first build our custom image, then login to Docker 
Hub, tag and push our image to the hub.

Next part is to create a Lowdefy front-end for our web service.