Complete expose our web app using cloudflared (#7)
This is part 7 of 7 parts of the end-to-end application that includes a
CI pipeline.

Currently, I'm hosting the Docker container Python Flask server app on
my local workstation, i.e. http://localhost:8080 and using cloudflared
as a service to access at https://archiveso.markit.work.

Next part is the cooldown period followed by the building cycle.