Complete Python Flask server authentication. (#9)
This is part 2 of 5 parts of the build cycle.

We are using HTTP basic authentication for our Python Flask server.
This is to prevent any unauthorized persons to access our API endpoints,
which are available via a public domain.

There are many methods of storing and retrieving secrets. However, in
this build we will only be covering the use of environment variables.

We need to ensure that the secrets are set in the environment variables on:

- a Python flask server when running locally
- a Docker container server when running locally
- a Docker container server when running in our CI pipeline
- a Docker container server when running in our production instance

Next part is to build a Lowdefy page to allow user interaction.