Complete build and test an API endpoint
This is part 4 of 4 parts of the build cycle.

When you approach coding using a test driven development ["TDD"], you
start with a test case and then build your code around the test case.
You can have have one or more test cases for each feature.

In our example, we are building an API endpoint that allows users to
add URL(s) with the `POST` method. There are two cases that we test for:

- Archivebox index not found
- Empty user form

In our first test case, with reference to our connection where users
are taken to, this fails if the `path/to/archivebox` is not configured
as an environment variable `ARCHIVEBOX_PATH`.

The second test case, with reference to our places where users can
navigate, this fails if the user form is empty and the body of the
`POST` method does not contain the `url` field.